make sure tests pass - refactor psth in broken parts of code
