+++ src/iblphotometry/pipelines.py
@@ -3,7 +3,8 @@
 import numpy as np
 import pandas as pd
 import pynapple as nap
-from iblphotometry.helpers import z, psth, filt
+from iblphotometry.helpers import z, filt
+from iblphotometry.behavior import psth
 
 # from ibldsp.utils import WindowGenerator
 from iblphotometry import sliding_operations
